+++ Parciales/Final2019/Pagina.cpp
@@ -17,7 +17,7 @@ Pagina::Pagina ()
 /**
  * Inicializa los valores con los datos pasados por parametro. (Constructor 2)
  */
-Pagina::Pagina (string d, uint8_t o, uint8_t dest, uint8_t idpag)
+Pagina::Pagina (string d, uint16_t o, uint16_t dest, uint8_t idpag)
 {
   this->dato = d;
   this->Origen = o;
@@ -38,7 +38,7 @@ string Pagina::getDato ()
  * Devuelve el Ip de origen, que es de la maquina que fue enviada la pagina
  * @return  Origen de tipo unsigned int de 8 bits.
  */
-uint8_t Pagina::getOrigen ()
+uint16_t Pagina::getOrigen ()
 {
   return Origen;
 }
@@ -47,7 +47,7 @@ uint8_t Pagina::getOrigen ()
  * Devuelve el Ip de destino almacenado en la pagina a enviar.
  * @return Destino de tipo unsigned int de 8 bits.
  */
-uint8_t Pagina::getDestino ()
+uint16_t Pagina::getDestino ()
 {
   return Destino;
 }
