+++ Parciales/Final2019/Pagina.h
@@ -12,15 +12,15 @@ using namespace std;
 class Pagina {
  private:
   string dato;
-  uint8_t Origen;
-  uint8_t Destino;
+  uint16_t Origen;
+  uint16_t Destino;
   uint8_t IDpag;
 
  public:
   Pagina ();
-  Pagina (string, uint8_t, uint8_t, uint8_t);
-  uint8_t getOrigen ();
-  uint8_t getDestino ();
+  Pagina (string, uint16_t, uint16_t, uint8_t);
+  uint16_t getOrigen ();
+  uint16_t getDestino ();
   string getDato ();
   uint8_t getIDpag();
